Tank move with terrain regen

+++ src/coordinate.h
@@ -5,9 +5,9 @@
 
 typedef struct {
   // between 0 to 319
-  uint16_t x;
+  int x;
   // between 0 to 239
-  uint16_t y;
+  int y;
 } Coordinate;
 
 void updateCoordinate(Coordinate *c, uint16_t newX, uint16_t newY);
